+++ federation/hostmeta/django/generators.py
@@ -1,8 +1,11 @@
 import logging
 
+# noinspection PyPackageRequirements
 from django.http import HttpResponseBadRequest, JsonResponse, HttpResponseNotFound
 
-from federation.hostmeta.generators import RFC7033Webfinger, generate_nodeinfo2_document
+from federation.hostmeta.generators import (
+    RFC7033Webfinger, generate_nodeinfo2_document, MatrixClientWellKnown, MatrixServerWellKnown,
+)
 from federation.utils.django import get_configuration, get_function_from_config
 from federation.utils.text import get_path_from_url
 
@@ -17,6 +20,34 @@ def nodeinfo2_view(request, *args, **kwargs):
     nodeinfo2 = nodeinfo2_func()
 
     return JsonResponse(generate_nodeinfo2_document(**nodeinfo2))
+
+
+def matrix_client_wellknown_view(request, *args, **kwargs):
+    try:
+        matrix_config_func = get_function_from_config("matrix_config_function")
+    except AttributeError:
+        return HttpResponseBadRequest("Not configured")
+    matrix_config = matrix_config_func()
+
+    wellknown = MatrixClientWellKnown(
+        homeserver_base_url=matrix_config["homeserver_base_url"],
+        identity_server_base_url=matrix_config.get("identity_server_base_url"),
+        other_keys=matrix_config.get("client_wellknown_other_keys"),
+    )
+    return JsonResponse(wellknown.render())
+
+
+def matrix_server_wellknown_view(request, *args, **kwargs):
+    try:
+        matrix_config_func = get_function_from_config("matrix_config_function")
+    except AttributeError:
+        return HttpResponseBadRequest("Not configured")
+    matrix_config = matrix_config_func()
+
+    wellknown = MatrixServerWellKnown(
+        homeserver_domain_with_port=matrix_config["homeserver_domain_with_port"],
+    )
+    return JsonResponse(wellknown.render())
 
 
 def rfc7033_webfinger_view(request, *args, **kwargs):