+++ federation/hostmeta/generators.py
@@ -3,6 +3,7 @@ import os
 import warnings
 from base64 import b64encode
 from string import Template
+from typing import Dict
 
 from jsonschema import validate
 from jsonschema.exceptions import ValidationError
@@ -329,6 +330,43 @@ def get_nodeinfo_well_known_document(url, document_path=None):
             }
         ]
     }
+
+
+class MatrixClientWellKnown:
+    """
+    Matrix Client well-known as per https://matrix.org/docs/spec/client_server/r0.6.1#server-discovery
+    """
+    def __init__(self, homeserver_base_url: str, identity_server_base_url: str = None, other_keys: Dict = None):
+        self.homeserver_base_url = homeserver_base_url
+        self.identity_server_base_url = identity_server_base_url
+        self.other_keys = other_keys
+
+    def render(self):
+        doc = {
+            "m.homeserver": {
+                "base_url": self.homeserver_base_url,
+            }
+        }
+        if self.identity_server_base_url:
+            doc["m.identity_server"] = {
+                "base_url": self.identity_server_base_url,
+            }
+        if self.other_keys:
+            doc.update(self.other_keys)
+        return doc
+
+
+class MatrixServerWellKnown:
+    """
+    Matrix Server well-known as per https://matrix.org/docs/spec/server_server/r0.1.4#server-discovery
+    """
+    def __init__(self, homeserver_domain_with_port: str):
+        self.homeserver_domain_with_port = homeserver_domain_with_port
+
+    def render(self):
+        return {
+            "m.server": self.homeserver_domain_with_port,
+        }
 
 
 class RFC7033Webfinger: